Fixing and streamlining bugs around flow verification
